requirements.txt
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.3.0
google-generativeai==0.3.0
python-dotenv==1.0.0
python-dateutil==2.9.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
pytest==8.3.4
